** Hotfix ** UI health bar text now only displays integer values. No more floating points when hit by explosions, etc.

+++ Assets/Scripts/UI/HealthBarPlayer.cs
@@ -49,7 +49,7 @@ public class HealthBarPlayer : MonoBehaviour {
 			if (text)
 			{
 				// Update text
-				text.text = (healthComponent.CurrentHealthPct.ToString());
+				text.text = (int)(healthComponent.CurrentHealthPct.ToString());
 			}
 		}
 	}
